Move xml-crypto and xmlenc operations to xml module
